Aaaaa

+++ Oefening 9/public_html/admin/edit.php
@@ -7,7 +7,7 @@ if (!(isset($_SESSION["admin"]) && $_SESSION["admin"] == 1)) {
     return;
 }
 
-if (!(isset($_POST["edit"]) || isset($_POST["delete"]) || isset($_POST["confirm-edit"]))) {
+if (!(isset($_POST["edit"]) || isset($_POST["delete"]) || isset($_POST["new-product"]) || isset($_POST["confirm-edit"]) || isset($_POST["confirm-new-product"]))) {
     header("Location: ./");
     return;
 }
@@ -26,9 +26,10 @@ if (!(isset($_POST["edit"]) || isset($_POST["delete"]) || isset($_POST["confirm-
     global $connection;
 
     $type = $_GET["type"];
-    $product = $_GET["product"];
 
+    /* Start editing */
     if (isset($_POST["confirm-edit"])) {
+        $product = $_GET["product"];
         $name = $_POST["name"];
         $price = $_POST["price"];
         $image = $_POST["image"];
@@ -46,11 +47,12 @@ if (!(isset($_POST["edit"]) || isset($_POST["delete"]) || isset($_POST["confirm-
         $statement->bind_param("sisiii", $name, $price, $image, $category, $platform, $product);
         $statement->execute();
 
-        header("Location: ./");
+        header("Location: ./products.php");
         exit();
     }
 
     if ($type == "edit") {
+        $product = $_GET["product"];
         $sql = "SELECT * FROM products WHERE id=?";
         $statement = $connection->prepare($sql);
 
@@ -74,13 +76,59 @@ if (!(isset($_POST["edit"]) || isset($_POST["delete"]) || isset($_POST["confirm-
                 <button name="confirm-edit">Edit</button>
             </form>
         ';
-    } else if ($type == "delete") {
+        exit();
+    }
+    /* End editing */
+
+    /* Start creation */
+    if (isset($_POST["confirm-new-product"])) {
+        $product = $_GET["product"];
+        $name = $_POST["name"];
+        $price = $_POST["price"];
+        $image = $_POST["image"];
+        $category = $_POST["category"];
+        $platform = $_POST["platform"];
+
+        $sql = "INSERT INTO products (name, price, image, category, platform) VALUES (?, ?, ?, ?, ?);";
+        
+        $statement = $connection->prepare($sql);
+
+        if (!$statement) {
+            die("Error: " . $connection->error);
+        }
+
+        $statement->bind_param("sisii", $name, $price, $image, $category, $platform);
+        $statement->execute();
+
+        header("Location: ./products.php");
+        exit();
+    }
+
+    if ($type == "create") {
+        echo '
+            <form action="./edit.php?type=create" method="post">
+                <input type="text" name="name" placeholder="Name">
+                <input type="text" name="price" placeholder="Price">
+                <input type="text" name="image" placeholder="Image">
+                <input type="text" name="category" placeholder="Category">
+                <input type="text" name="platform" placeholder="Platform">
+                <button name="confirm-new-product">Create</button>
+            </form>
+        ';
+
+        exit();
+    }
+    /* End creation */
+    
+    if ($type == "delete") {
+        $product = $_GET["product"];
         $sql = "DELETE FROM products WHERE id=?";
         $statement = $connection->prepare($sql);
         $statement->bind_param("i", $product);
         $statement->execute();
 
         header("Location: ./");
+        exit();
     }
     ?>
 </body>
